pyproject.toml
[project]
name = "nerion-selfcoder"
version = "0.0.1"
description = "Nerion Selfcoder toolkit"
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Nerion Team" }]
keywords = ["ast", "autocoder", "refactoring", "developer-tools"]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Build Tools",
  "Topic :: Software Development :: Code Generators",
]
dependencies = [
  # Minimal runtime deps required by CLI and core modules
  "PyYAML>=6.0",
  "pydantic>=2.1,<3",
  "requests>=2.32",
]

[project.urls]
Homepage = "https://example.com/nerion"
Repository = "https://example.com/nerion/repo"
Issues = "https://example.com/nerion/issues"

[project.scripts]
nerion = "selfcoder.cli:main"
selfcoder = "selfcoder.cli:main"
nerion-chat = "app.nerion_chat:main"
nerion-health = "selfcoder.cli.health:main"

[project.optional-dependencies]

dev = [
  "pytest>=7",
  "ruff>=0.3",
]

# PDFs only
docs-pdf = [
  "pypdf>=4.0.0",
]

# URLs: plain fetch + heavy JS rendering
docs-web = [
  "requests>=2.32",
  "playwright>=1.45",
  "trafilatura>=1.8",
  "beautifulsoup4>=4.12",
]

[tool.ruff]
line-length = 100
target-version = "py39"
extend-exclude = [
  "legacy_import",
  # Exclude tests from strict lint to keep signal focused on library/runtime code
  "tests",
  "selfcoder/tests",
  # Demo/fixture code intentionally noisy
  "src/demo.py",
]

[tool.ruff.lint.per-file-ignores]
# Voice stack allows imports after docstrings/future imports; quiet E402 noise
"voice/stt/recognizer.py" = ["E402"]
"voice/stt/vad.py" = ["E402"]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["selfcoder*", "app*"]
